+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased] - TBD
+### Added
+- `GetEventSubscriptionsListAsync` method with enhanced filtering and sorting capabilities for retrieving event subscriptions
+
 ### Changed
 - Upgraded .NET Core runtime to .NET 7.0 for Tests and Examples projects
 - Drop support for .NET Core 2.1 and 3.1 (dropper netstandard 1.x)